requirements.txt
streamlit
openpyxl
langchain_community==0.3.11
python-dotenv
langchain==0.3.11
pandas==2.2.2
sqlalchemy==2.0.36
pymysql
seaborn==0.13.2
matplotlib
boto3
awscli
langchain_aws
flask
flask_cors
numpy
crewai
crewai_tools